requirements.txt
base58==2.0.1
certifi==2020.11.8
chardet==3.0.4
idna==2.10
ipfshttpclient==0.7.0a1
multiaddr==0.0.9
netaddr==0.8.0
pycryptodome==3.9.9
requests==2.25.0
six==1.15.0
urllib3==1.26.2
varint==1.0.2
